loading and saving model parameters using gguf format

+++ code/src/Constants.cs
@@ -2,11 +2,11 @@
 
 public static class Constants
 {
-    public const double ZERO = 0;
+    public const float ZERO = 0f;
     public const int ONE = 1;
-    
-    public const double GRADIENT_PLUS = 1;
-    public const double GRADIENT_MINUS = -1;
+
+    public const float GRADIENT_PLUS = 1f;
+    public const float GRADIENT_MINUS = -1f;
     
     public const string WEIGHT_PREFIX = "w";
     public const string BIAS_PREFIX = "b";
