+++ code/use-cases/biagram/SamplingUtils.cs
@@ -3,12 +3,12 @@ using TinyBrain;
 
 public static class SamplingUtils
 {
-    public static int Multinomial(double[] probabilities)
+    public static int Multinomial(float[] probabilities)
     {
-        var total = 0.0;
+        var total = 0.0f;
         for (var i = 0; i < probabilities.Length; i++) total += probabilities[i];
-        var r = Random.Shared.NextDouble() * total;
-        var cumulative = 0.0;
+        var r = (float)Random.Shared.NextDouble() * total;
+        var cumulative = 0.0f;
         for (var i = 0; i < probabilities.Length; i++)
         {
             cumulative += probabilities[i];
@@ -20,9 +20,9 @@ public static class SamplingUtils
     // Returns Operand [N, numClasses] — each row is a one-hot vector
     public static Operand OneHotMatrix(int[] indices, int numClasses)
     {
-        var data = new double[indices.Length, numClasses];
+        var data = new float[indices.Length, numClasses];
         for (var i = 0; i < indices.Length; i++)
-            data[i, indices[i]] = 1.0;
+            data[i, indices[i]] = 1.0f;
         return Operand.Of(data);
     }
 }
